requirements.txt
django 
channels 
psutil
djangorestframework
djangorestframework-simplejwt
django-cors-headers
daphne 